==> 01_features_nsrr_cfs.ipynb <==
# NSRR CFS Features extraction

**WARNING:** 

- CFS v1 and v2 files have different sampling rates:
    - v1: 128 EEG + EOG, 256 ECG / EMG
    - v2: 128 EEG + EOG, 512 ECG and 256 EMG
- All EEG, EOG and EMG are referenced to Fpz.
- Highpass at 0.16 Hz.

https://sleepdata.org/datasets/cfs/pages/polysomnography-introduction.md

import os
import yasa
import platform
import warnings
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm
from mne.io import read_raw_edf

# Define paths
if platform.system() == 'Darwin':
    root_dir = '/Volumes/NSRR/cfs/'
else:
    root_dir = 'G:/cfs/'
eeg_dir = root_dir + 'polysomnography/edfs/'
hypno_dir = root_dir + 'polysomnography/annotations-events-profusion/'

desc_dir = root_dir + 'datasets/cfs-visit5-dataset-0.4.0.csv'

df_subj = pd.read_csv(desc_dir, 
    usecols=['nsrrid', 'age', 'SEX', 'race', 'ethnicity', 'bmi', 'ahi_a0h3',
             'AbnorEEG', 'AbnorEye', 'QuEEG1', 'QuHR', 'QuEOGL', 'QuChin', 
             'RemNRemPr', 'Stg1Stg2Pr', 'Stg2Stg3Pr', 'WakSlePr'])

# Rename for ease-of-use
df_subj.rename(columns={
    'nsrrid': 'subj',
    'SEX': 'male',
    'ahi_a0h3': 'ahi',
}, inplace=True)

df_subj['race'].replace({1: 'caucasian', 2: 'african', 3: 'other'}, inplace=True)
df_subj.loc[df_subj['ethnicity'] == 1, 'race'] = 'hispanic'
df_subj.drop(columns=['ethnicity'], inplace=True)
df_subj.rename(columns={'race': 'ethnicity'}, inplace=True)

# Remove subjects with no EEG data file
df_subj = df_subj[~df_subj['subj'].isin([800269, 802230])]

# Remove Ss with bad data quality or unreliable scoring
df_subj = df_subj[df_subj['RemNRemPr'] != 1]
df_subj = df_subj[df_subj['Stg1Stg2Pr'] != 1]
df_subj = df_subj[df_subj['Stg2Stg3Pr'] != 1]
df_subj = df_subj[df_subj['WakSlePr'] != 1]
df_subj = df_subj[df_subj['AbnorEEG'] != 1]
df_subj = df_subj[df_subj['AbnorEye'] != 1]
df_subj = df_subj[df_subj['QuEEG1'] >= 4] # >= 4 = signal good for > 75% of the time
df_subj = df_subj[df_subj['QuEOGL'] >= 4]
df_subj = df_subj[df_subj['QuChin'] >= 4]

# Drop columns
df_subj.drop(columns=['RemNRemPr', 'AbnorEEG', 'AbnorEye', 'QuEEG1', 'QuHR', 'QuEOGL', 'QuChin', 
                      'RemNRemPr', 'Stg1Stg2Pr', 'Stg2Stg3Pr', 'WakSlePr'], inplace=True)

# Convert to str
df_subj['subj'] = df_subj['subj'].astype(str)
df_subj.set_index('subj', inplace=True)

# Export demographics to CSV file
df_subj['dataset'] = 'CFS'
df_subj.to_csv("/Volumes/JAWA/SHERLOCK/demo/demo_nsrr_cfs.csv")

# Keep only a random subset of 550 subjects (leave ~100 for testing)
df_subj = df_subj.sample(n=550, replace=False, random_state=42)

print(df_subj.shape[0], 'subjects remaining')
df_subj.head(10)

df = []
include = ['C4', 'LOC', 'EMG1']
sf = 100

for sub in tqdm(df_subj.index):
    eeg_file = eeg_dir + 'cfs-visit5-' + str(sub) + '.edf'
    hypno_file = hypno_dir + 'cfs-visit5-' + str(sub) + '-profusion.xml'

    # LOAD EEG DATA
    try:
        raw = read_raw_edf(eeg_file, preload=False, verbose=0)
        raw = read_raw_edf(eeg_file, preload=True, 
                           exclude=np.setdiff1d(raw.info['ch_names'], include), 
                           verbose=0)
    except:
        continue
        
    # Resample and low-pass filter
    raw.resample(sf, npad="auto")
    
    # LOAD HYPNOGRAM
    hypno, sf_hyp = yasa.load_profusion_hypno(hypno_file)
    # We keep up to 15 minutes before / after sleep
    start_to_firstsleep_min = np.nonzero(hypno)[0][0] / 2
    lastsleep = np.nonzero(hypno)[0][-1]
    lastsleep_to_end_min = (len(hypno) - lastsleep) / 2
    tmin, tmax = 0, None  # must be in seconds
    if start_to_firstsleep_min > 15:
        tmin = (start_to_firstsleep_min - 15) * 60
    if lastsleep_to_end_min > 15:
        tmax = lastsleep * 30 + 15 * 60
    # Crop!
    raw.crop(tmin, tmax)
    if tmax is None:
        hypno = hypno[int(tmin / 60 * 2):]
    else:
        hypno = hypno[int(tmin / 60 * 2):int(tmax / 60 * 2)]
    # Hypno and data have the same number of epochs
    n_epochs = hypno.shape[0]
    if n_epochs != np.floor(raw.n_times / sf / 30):
        print("- Hypno and data size do not match.")
        continue
    
    # Convert hypnogram to str
    df_hypno = pd.Series(hypno)
    df_hypno.replace({0: 'W', 1: 'N1', 2: 'N2', 3: 'N3', 4: 'R'}, inplace=True)
    stage_min = df_hypno.value_counts(sort=False) / 2

    # INCLUSION CRITERIA
    # Hypnogram must include all stages
    if np.unique(hypno).tolist() != [0, 1, 2, 3, 4]:
        print("- Not all stages are present.")
        continue
    # If the duration is not between 4 to 12 hours, skip subject
    if not(4 < n_epochs / 120 < 12):
        print("- Recording too short/long.")
        continue
    # Requires at least 5 min of each stage
    # if (stage_min < 5).any():
    #    print("- Not 5 min of each stage.")
    #    continue
       
    # EXTRACT FEATURES
    metadata = dict(age=df_subj.loc[sub, 'age'], male=df_subj.loc[sub, 'male'])
    sls = yasa.SleepStaging(raw, eeg_name=include[0], eog_name=include[1], 
                            emg_name=include[2], metadata=metadata)

    features = sls.get_features().reset_index()
    features['subj'] = sub
    features.set_index(['subj', 'epoch'], inplace=True)
    
    # Add hypnogram
    features['stage'] = df_hypno.to_numpy()
    df.append(features)

df = pd.concat(df)

# Add dataset
df['dataset'] = 'cfs'

# Convert to category
df['dataset'] = df['dataset'].astype('category')
df['stage'] = df['stage'].astype('category')

# Show %stage
df['stage'].value_counts(normalize=True, sort=True)

# Median value of the EEG IQR per stage
df.groupby('stage')['eeg_iqr'].median()

# Number of unique nights in dataset
df.index.get_level_values(0).nunique()

# Export feature file
if platform.system() == 'Darwin':
    df.to_parquet("/Volumes/JAWA/SHERLOCK/features/features_nsrr_cfs.parquet")
else:
    df.to_parquet("H:/SHERLOCK/features/features_nsrr_cfs.parquet")
